OLPC 1.5: Host reboot no longer watchdog reboots the EC.
In order to unifiy the reboot code paths 1.5 reboots were changed such that
they rebooted the EC as well.  This cleared the indexed IO bit so things
like flash upgrded didn't need a 2nd reboot.
However, We really don't want to be rebooting the EC so much since
the battery subsystem does not have any sort of sync to shutdown
prior to the reboot.
So rather than use cmd66 0xdb cmd there is a new cmd66 0xd7 which will
power cycle the host but not reboot the EC.  Index IO is now enabled
after the EC drops power to the host so it achieves the same behavior as 0xdb.
0xdb is now only for firmware updates.

git-svn-id: svn://openfirmware.info/openfirmware@1664 1552c027-8020-0410-b4b5-a757f869b4ce