scsidisk - retry TEST UNIT READY once
Some USB flash drives violate section 3.4 of the USB Mass Storage
Specification For Bootability 1.0, October 25, 2004, by responding to
TEST UNIT READY with CHECK CONDITION status with a sense key of NOT
READY, and an additional sense code of MEDIUM NOT PRESENT.

Section 3.4 says "If the Device cannot become operational or in a
state such that a Host Computer action is required to make the Device
ready, the Device returns CHECK CONDITION status with a sense key of
NOT READY."

So the device either cannot become operational, or requires action of
the host.

However, a second TEST UNIT READY is successful, without any other
action of the host other than trying again.

READ CAPACITY and READ commands also fail if they are done first,
before the device admits to being ready.

Section 3.4 also says "Devices are expected to respond promptly to
indicate the current status of the Device."

Section 6.1 of the USB Mass Storage Class - Compliance Test
Specification 0.9a, June 30, 2006, says "Issue a Test Unit Ready
command.  Repeat until the device indicates a ready state."



git-svn-id: svn://openfirmware.info/openfirmware@3726 1552c027-8020-0410-b4b5-a757f869b4ce