mips: fix the build
Apart from fixing a couple of minor issues, and dropping some artifacts,
this adds a Makefile in cpu/mips/build/, based on what the other
architectures do.

It also fixes the botched whitespace in Makefile in cpu/mips/Linux/ and
switches the default compiler to a slightly different one. This way I'm
able to build the builder.dic on a recent Fedora x86 system.