mips/sys: allow emulating the syscalls
This issues a syscall instruction if the syscall vector is empty, much
like other architectures do. This allows an emulator to sevice the call.

Not being too familiar with good MIPS assembly practices, I suspect there
might be a better way to do this. This works well enough for now.