video/textmode: set O/E access when setting the font
The alternating Odd/Even mapping of VGA memory accesses needs to be off
for writing the font to plane 2 and on when accessing the text/attribute
data in planes 0 and 1. This is done with the bit 3 (Host O/E) of register 5
(Graphics Mode Register).

This unbreak font setting on QEMU emulated VGA card.

Tested not to break ega-output on XO-1. Not sure why it worked; perhaps
the Geode graphics makes some assumptions from what's set in register 6
or something.