egatext: optionally offset text fb base by ISA memory base
Currently the ega-text node is typically placed under the root node and
thus its attempt to map in the text framebuffer assumes it's mapped
directly into the CPU's memory space.

In fact, the VGA card might be attached to an ISA bus bridge and mapped
somewhere else in the address space. This is the case on machines where
the physical memory resides toward the bottom of the address space and
devices are mapped elsewhere, such as on MIPS machine (as emulated by
QEMU, but perhaps physical ones as well).

Let's deal with it by offsetting the text frame buffer space by an
optional constant, if defined by a board.

A proper fix for this would probably involve merging this into the VGA
device node or reparenting this below it, but I suppose this will do for
now.