termemu: add cpu independent fb8-ops.fth
Provide machine independent implementations of some words that are
generally used around termemu and expected to be provides by cpu-specific
parts of the tree.

They are neither pretty nor fast and generally just serve as placeholders
until fast versions are implemented in assembly.

The assembly implementations are notably missing on MIPS.