video/textmode: load font data to 0xa0000 not 0xb8000
The page 2 with font data is usually mapped at the beginning of the
graphics memory at 0xa0000 whereas pages 0 and 1 are interleaved at
0xb8000.

Perhaps the real hardware doesn't enforce it, but the QEMU emulated VGA
adapter seems to.

Tested not to break on real hardware by issuing "vga-output" on a XO-1.