video/textmode: allow mapping VGA memory on ISA cards
The VGA memory space calls parent bus' mapping routine with arguments
that are tailored towards the PCI bus.

This will not work on platforms with an ISA VGA card, such as the MIPS
machine emulated by QEMU.

The said QEMU MIPS machine will define "isa-mem-base", so let's assume
we want the ISA address there.

I guess the correct way to deal with this would be to get the proper
address using "#address-cells" and "reg" properties but I have not idea
how to do that.